Produc loja

+++ loja/produtos/forms.py
@@ -1,10 +1,10 @@
 from flask_wtf.file import FileAllowed, FileField, FileRequired
-from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators
+from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators, DecimalField
 
 
 class Addprodutos(Form):
     name = StringField('Nome :', [validators.DataRequired()])
-    price = IntegerField('Preço :', [validators.DataRequired()])
+    price = DecimalField('Preço :', [validators.DataRequired()])
     discount = IntegerField('Desconto :', [validators.DataRequired()])
     stock = IntegerField('Estoque :', [validators.DataRequired()])
     discription = TextAreaField('Descrição :', [validators.DataRequired()])
